Port code to use new Libica-2.x API.
Introduce necessary changes to the IBMCA OpenSSL Dynamic engine code
to use the new API provided by the recently released Libica version 2.

This also removes the ica_openssl_api.h header file, which was
just a local version for the ica_api.h, and includes minor cleanups
to the configure script and Makefiles.

Signed-off-by: Klaus Heinrich Kiwi <klausk@linux.vnet.ibm.com>