Fix wrong blocksize handling with stream cipher modes.
Corrected wrong reporting of blocksize value for DES, 3DES and
AES for ofb and cfb mode. The blocksize value to be reported via
EVP_CIPHER struct to openssl needs to be 1 for stream cipher
mode to process and en/decrypt the exact amount of input data
bytes into the very same amount of output data bytes. Code has
been tested and verified.

Signed-off-by: Harald Freudenberger <freude@linux.vnet.ibm.com>
Signed-off-by: Claudio Carvalho <cclaudio@linux.vnet.ibm.com>