Make sure ibmca_init only runs once.
The ibmca_init function is called twice or maybe even more.
However, with some printf added it is clearly to see that
the init is run 2 times obviously there is a race condition
with the static init variable within the ibmca_init function.
With this patch the variable is incremented at the very
beginning of the function. There is still a small gap open
for a race condition but with tests I have not been able
to hit this and now ibmca (and so Libica) is initialized
only once.

Signed-off-by: Harald Freudenberger <freude@linux.vnet.ibm.com>