Dont fail when a libica symbol cannot be resolved.
The engines init/finish function was used as constructor/destructor.
Engine init/finish is called per engine object, so move all code
that is unrelated to initializing/uninitializing engine objects to
new contructor/destructor functions: The libica symbols will be
resolved in the constuctor and the engine will not fail if a
symbol cannot be resolved.

Signed-off-by: Patrick Steuer <patrick.steuer@de.ibm.com>