TESTS: OpenSSL 3.0 compatibility
DES is only available in the legacy provider in OpenSSL 3.0.  We skip these
tests for OpenSSL 3.0 since we cannot be sure the legacy provider is loaded
and do not want to use yet another configuration to ensure the legacy provider
is loaded.

For 3DES-ECB, OpenSSL 3.0 requires an iv (which it ignores) on the command
line.  Otherwise, execution fails which fails the test.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>