Update to libica.so.4
Add configure time switch --with-libica-version to specify the major version
of libica the engine should use by default.

Update enginectrl test to test for both libica version 3 and 4.  It
automatically skips the libraries that are not available on the system.  So
testing for both is not a problem and only produces some information logs for
all non-existing libraries.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>