Fix compilation for OpenSSL 3.0
ECX_KEY structure as used by engine is not public API in OpenSSL 3.0 and
should not be included via a private header.  This forced the engine to only
compile if the OpenSSL 3.0 source tree is present.  While this might be the
cleaner solution, we now copy out the defines to be used.  This makes the
engine compile even without OpenSSL 3.0 source tree.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>