tests: skip tests if libica does not support required algorithms
Before actually running the tests, check if libica supports the
required algorithms. Skip the whole test if not.

This can happen when running the test on a system without appropriate
crypto adapters. This would lead to the situation that the provider would
not register itself for the required algorithms, and thus the OpenSSL
default provider would be used. This would make the tests to fail, because
it is not running with the IBMCA provider as expected by the test.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>