provider: Fix retrieval of non-existing private key params
When sensitive private key parameters are retrieved via ibmca_param_get_bn(),
a secure BigNum is allocated via BN_secure_new() before trying to get the
parameter from the params array. If the desired parameter is not available
(e.g. because it is optional), then the previously allocated BigNum must be
freed and set to NULL.

Leaving the BigNum allocated but zero, may cause trouble later on, when it is
tried to work with such a private key part. When exporting such a zero private
key part and importing it into another provider, this may for example fail
because this may result in a zero length integer OSSL_PARAM, i.e. an OSSL_PARAM
with a NULL pointer as data. Other provides do not expect such empty parameters,
and may fail or segfault.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>