provider: RSA: Fix get_params to retrieve max-size, bits, and security-bits
The RSA key management's get_params() function should be able to return the
values for max-size, bits, and security-bits if at least the public key is
available.

The detection whether the key is 'empty', i.e. has neither the public nor the
private key components was wrong. This leads to the fact that those parameters
were not returned when only the public key was available.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>